+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## v1.1.2
+
+* New action `issue.create`
+
+## v1.1.1
+
+* Fix - Remove default group id
+
+## v1.1.0
+
+* Use of `python-gitlab` library
+* New action `epic.create`
+
 ## v1.0.1
 
 * Small bug fixes regarding Python 3 support